两根一剔除

+++ main.py
@@ -2,7 +2,6 @@
 
 import multiprocessing
 import os
-import shutil
 import sys
 from datetime import datetime
 from PySide6.QtWidgets import QApplication
@@ -13,8 +12,8 @@ from qt_material import apply_stylesheet
 if __name__ == "__main__":
     multiprocessing.freeze_support()
 
-    # if not os.path.exists("C:/Windows/software.config"):
-    #     exit(0)
+    if not os.path.exists("C:/Windows/software.config"):
+        exit(0)
     if os.path.isfile("log.txt"):
         os.remove("log.txt")
     # if os.path.isdir("./All_Images"):
